// src/app/models/User.ts
export interface User {
   id?: number;
   document?: number;
   name?: string;
   lastname?: string;
   email?: string;
   phone?: number;
   username?: string;
   password?: string;
   user_status?: boolean;
   exitoso?: string;
   mensajeEjecucion?: string;
   createdAt?: Date;
   updatedAt?: Date;
   _id?: number;
   __v?: string;
}